test(coverage): move parser-only assertions out of contracts directory
CodeRabbit flagged that the extractCoverageMeasurementId describe block
in coverage.test.ts (under tests/contracts/) bypasses the
ContractScenario / ContractOperation harness — it's parser-only unit
testing, not contract testing. Moved the entire block to a new
tests/coverage-link.test.ts file outside tests/contracts/, keeping the
contracts directory purely structural.

Resolves CodeRabbit review comment on PR #175 (iteration 2).
